pico: Move all the display code out of main
It was the last big thing in there
